+++ Libraries/SideScroll.Tabs.Samples/DataGrid/TabSampleDataGridHashSet.cs
@@ -12,7 +12,7 @@ public class TabSampleDataGridHashSet : ITab
 		{
 			_items = [];
 			AddEntries();
-			model.AddData(_items);
+			model.AddItems(_items);
 
 			// HashSet not observable
 			/*model.Actions = new List<TaskCreator>